refactor IQ into common_utils
