+++ qualibration_graphs/quantum_dots/calibration_utils/psb_search_sweep_detuning/simulated_data_generator.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING, List, Optional, Sequence, Union
 import numpy as np
 import xarray as xr
 
-from calibration_utils.iq_utils.iq_blobs.readout_barthel.simulate import (
+from calibration_utils.common_utils.iq_utils.iq_blobs.readout_barthel.simulate import (
     SimulationParamsIQ,
     simulate_readout_iq,
 )
