+++ qualibration_graphs/quantum_dots/calibration_utils/psb_search_sweep_detuning_vs_buffer/parameters.py
@@ -4,7 +4,7 @@ from qualibrate.core import NodeParameters
 from qualibrate.core.parameters import RunnableParameters
 from qualibration_libs.parameters import CommonNodeParameters, QubitPairExperimentNodeParameters
 
-from calibration_utils.iq_utils import IQSweepParameters
+from calibration_utils.common_utils.iq_utils import IQSweepParameters
 
 
 class NodeSpecificParameters(RunnableParameters):
